ftPp splitting and matching
