+++ asm/melee/ft/chara/ftNn_Init.s
@@ -1146,7 +1146,7 @@ ftPp_SpecialS_0_Coll:
 /* 80123FC8 00120BA8  2C 03 00 00 */	cmpwi r3, 0
 /* 80123FCC 00120BAC  40 82 00 78 */	bne .L_80124044
 /* 80123FD0 00120BB0  7F 63 DB 78 */	mr r3, r27
-/* 80123FD4 00120BB4  4B FF B7 29 */	bl ftPp_SpecialN_8011F6FC
+/* 80123FD4 00120BB4  4B FF B7 29 */	bl ftPp_SpecialS_8011F6FC
 /* 80123FD8 00120BB8  2C 03 00 00 */	cmpwi r3, 0
 /* 80123FDC 00120BBC  40 82 00 68 */	bne .L_80124044
 /* 80123FE0 00120BC0  80 BE 00 B0 */	lwz r5, 0xb0(r30)
@@ -1282,7 +1282,7 @@ ftPp_SpecialS_1_Coll:
 /* 801241D0 00120DB0  2C 03 00 00 */	cmpwi r3, 0
 /* 801241D4 00120DB4  40 82 00 78 */	bne .L_8012424C
 /* 801241D8 00120DB8  7F 63 DB 78 */	mr r3, r27
-/* 801241DC 00120DBC  4B FF B5 21 */	bl ftPp_SpecialN_8011F6FC
+/* 801241DC 00120DBC  4B FF B5 21 */	bl ftPp_SpecialS_8011F6FC
 /* 801241E0 00120DC0  2C 03 00 00 */	cmpwi r3, 0
 /* 801241E4 00120DC4  40 82 00 68 */	bne .L_8012424C
 /* 801241E8 00120DC8  80 BE 00 B0 */	lwz r5, 0xb0(r30)
