+++ src/melee/ft/chara/ftPopo/ftPp_SpecialN.h
@@ -1,0 +1,21 @@
+#ifndef _FTPP_SPECIALN_
+#define _FTPP_SPECIALN_
+
+#include "ftPopo/forward.h"
+#include <baselib/forward.h>
+
+#include "ft/types.h"
+
+/* 11F1F8 */ void ftPp_SpecialN_Enter(HSD_GObj* gobj);
+/* 11F26C */ void ftPp_SpecialAirN_Enter(HSD_GObj* gobj);
+/* 11F318 */ void ftPp_SpecialN_Anim(HSD_GObj* gobj);
+/* 11F354 */ void ftPp_SpecialAirN_Anim(HSD_GObj* gobj);
+/* 11F390 */ void ftPp_SpecialN_IASA(HSD_GObj* gobj);
+/* 11F394 */ void ftPp_SpecialAirN_IASA(HSD_GObj* gobj);
+/* 11F398 */ void ftPp_SpecialN_Phys(HSD_GObj* gobj);
+/* 11F3B8 */ void ftPp_SpecialAirN_Phys(HSD_GObj* gobj);
+/* 11F3D8 */ void ftPp_SpecialN_Coll(HSD_GObj* gobj);
+/* 11F454 */ void ftPp_SpecialAirN_Coll(HSD_GObj* gobj);
+/* 11F500 */ void ftPp_SpecialN_8011F500(HSD_GObj* gobj);
+
+#endif
