+++ src/melee/ft/chara/ftPopo/ftPp_SpecialS.h
@@ -1,8 +1,15 @@
 #ifndef GALE01_11F99C
 #define GALE01_11F99C
 
+#include "ftPopo/forward.h"
 #include <baselib/forward.h>
 
+#include "ft/types.h"
+
+/* 11F68C */ void ftPp_SpecialS_8011F68C(HSD_GObj* gobj);
+/* 11F6FC */ bool ftPp_SpecialS_8011F6FC(HSD_GObj* gobj);
+/* 11F720 */ void ftPp_SpecialS_8011F720(HSD_GObj* gobj);
+/* 11F964 */ bool ftPp_SpecialS_8011F964(HSD_GObj*);
 /* 11F99C */ void ftPp_SpecialS_Enter(HSD_GObj* gobj);
 /* 11FB08 */ void ftPp_SpecialAirS_Enter(HSD_GObj* gobj);
 /* 11FC78 */ void ftPp_SpecialS1_Anim(HSD_GObj* gobj);
